content-security-policy/resource-hints/prefetch-allowed-no-default.html

<!DOCTYPE html>
  <html>
  <head>
    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline' img-src 'self' 'unsafe-inline' connect-src 'self' 'unsafe-inline' object-src 'self' 'unsafe-inline' font-src 'self' 'unsafe-inline' child-src 'self' 'unsafe-inline'">
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script src='/common/utils.js'></script>
    <script src='/content-security-policy/support/testharness-helper.js'></script>
    <script src='/content-security-policy/support/prefetch-helper.js'></script>
    <script>
      promise_test(async t => {
        assert_true(await try_to_prefetch("/common/dummy.xml", t));
      }, 'Prefetch should succeed when there is no default-src');
    </script>
  </head>
  <body>
  </body>
  </html>
